+++ resources/views/search_product/search_product.blade.php
@@ -221,9 +221,8 @@
 
                             var product_id = item.prd_id;
                             var price = parseFloat(item.price);
-                            var quantity = parseFloat(item.quantity);
-
-                            var subtotal = parseInt(price * quantity);
+                            var quantity = parseInt(item.quantity);
+                            var subtotal = parseFloat(price * quantity);
                             var cart_total = cart_total + subtotal;
                             var total_cart_quantity = total_cart_quantity + quantity;
                             $('#subtotal_' + product_id).html('$'+subtotal);
@@ -260,9 +259,8 @@
 
                             var product_id = item.prd_id;
                             var price = parseFloat(item.price);
-                            var quantity = parseFloat(item.quantity);
-
-                            var subtotal = parseInt(price * quantity);
+                            var quantity = parseInt(item.quantity);
+                            var subtotal = parseFloat(price * quantity);
                             var cart_total = cart_total + subtotal;
                             var total_cart_quantity = total_cart_quantity + quantity;
                             $('#subtotal_' + product_id).html('$'+subtotal);
@@ -408,9 +406,8 @@
 
                             var product_id = item.prd_id;
                             var price = parseFloat(item.price);
-                            var quantity = parseFloat(item.quantity);
-
-                            var subtotal = parseInt(price * quantity);
+                            var quantity = parseInt(item.quantity);
+                            var subtotal = parseFloat(price * quantity);
                             var cart_total = cart_total + subtotal;
                             var total_cart_quantity = total_cart_quantity + quantity;
                             $('#subtotal_' + product_id).html('$' + subtotal);
@@ -443,9 +440,8 @@
 
                             var product_id = item.prd_id;
                             var price = parseFloat(item.price);
-                            var quantity = parseFloat(item.quantity);
-
-                            var subtotal = parseInt(price * quantity);
+                            var quantity = parseInt(item.quantity);
+                            var subtotal = parseFloat(price * quantity);
                             var cart_total = cart_total + subtotal;
                             var total_cart_quantity = total_cart_quantity + quantity;
                             $('#subtotal_' + product_id).html('$' + subtotal);
